pyproject.toml
[tool.poetry]
name = "detect_seq"
version = "1.0.0"
description = ""
authors = ["MENG Haowei", "hermanzhaozzzz <hermanzhaozzzz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9.12"
_libgcc_mutex = "0.1"
_openmp_mutex = "4.5"
aioeasywebdav = "2.4.0"
aiohttp = "3.8.1"
aiosignal = "1.2.0"
amply = "0.1.5"
appdirs = "1.4.4"
asttokens = "2.0.5"
async-timeout = "4.0.1"
asynctest = "0.13.0"
attmap = "0.13.2"
attrs = "21.4.0"
backcall = "0.2.0"
bcrypt = "3.2.0"
bedtools = "2.30.0"
biopython = "1.78"
bismark = "0.23.1"
blas = "1.0"
blosc = "1.21.0"
boto3 = "1.24.28"
botocore = "1.27.28"
bottleneck = "1.3.5"
bowtie = "1.3.1"
bowtie2 = "2.4.5"
brotlipy = "0.7.0"
bwa = "0.7.17"
bzip2 = "1.0.8"
c-ares = "1.18.1"
ca-certificates = "2022.07.19"
cachetools = "4.2.2"
certifi = "2022.6.15"
cffi = "1.15.0"
charset-normalizer = "2.0.4"
coin-or-cbc = "2.10.8"
coin-or-cgl = "0.60.6"
coin-or-clp = "1.17.7"
coin-or-osi = "0.108.7"
coin-or-utils = "2.11.6"
coincbc = "2.10.8"
colorama = "0.4.5"
commonmark = "0.9.1"
configargparse = "1.4"
connection_pool = "0.0.3"
cryptography = "37.0.1"
cutadapt = "4.1"
dataclasses = "0.8"
datrie = "0.8.2"
dbus = "1.13.18"
decorator = "5.1.1"
defusedxml = "0.7.1"
dnaio = "0.9.0"
docutils = "0.18.1"
dpath = "2.0.6"
dropbox = "11.14.0"
executing = "0.8.3"
expat = "2.4.4"
fastqc = "0.11.9"
filechunkio = "1.8"
filelock = "3.6.0"
font-ttf-dejavu-sans-mono = "2.37"
fontconfig = "2.13.1"
freetype = "2.11.0"
frozenlist = "1.2.0"
ftputil = "5.0.4"
future = "0.18.2"
gitdb = "4.0.7"
gitpython = "3.1.18"
glib = "2.69.1"
google-api-core = "2.8.2"
google-api-python-client = "2.54.0"
google-auth = "2.6.0"
google-auth-httplib2 = "0.1.0"
google-cloud-core = "2.2.2"
google-cloud-storage = "1.43.0"
google-crc32c = "1.1.2"
google-resumable-media = "1.3.1"
googleapis-common-protos = "1.56.4"
grpcio = "1.42.0"
hdf5 = "1.12.1"
hisat2 = "2.2.1"
htslib = "1.15.1"
httplib2 = "0.20.4"
icu = "58.2"
idna = "3.3"
importlib-metadata = "4.11.3"
importlib_metadata = "4.11.3"
importlib_resources = "5.2.0"
iniconfig = "1.1.1"
ipython = "8.4.0"
isa-l = "2.30.0"
jedi = "0.18.1"
jinja2 = "3.0.3"
jmespath = "0.10.0"
jsonschema = "4.4.0"
jupyter_core = "4.10.0"
keyutils = "1.6.1"
krb5 = "1.19.3"
ld_impl_linux-64 = "2.38"
ldc = "1.28.1"
libblas = "3.9.0"
libcblas = "3.9.0"
libcrc32c = "1.1.1"
libcurl = "7.83.1"
libdeflate = "1.10"
libedit = "3.1.20210910"
libev = "4.33"
libffi = "3.3"
libgcc-ng = "12.1.0"
libgfortran-ng = "11.2.0"
libgfortran5 = "11.2.0"
libgomp = "12.1.0"
liblapack = "3.9.0"
liblapacke = "3.9.0"
libllvm12 = "12.0.1"
libnghttp2 = "1.47.0"
libopenblas = "0.3.20"
libpng = "1.6.37"
libprotobuf = "3.20.1"
libsodium = "1.0.18"
libssh2 = "1.10.0"
libstdcxx-ng = "12.1.0"
libuuid = "1.0.3"
libxcb = "1.15"
libxml2 = "2.9.14"
libzlib = "1.2.12"
logmuse = "0.2.6"
lz4-c = "1.9.3"
lzo = "2.10"
markupsafe = "2.1.1"
matplotlib-inline = "0.1.2"
mock = "4.0.3"
multidict = "5.2.0"
nbformat = "5.3.0"
ncurses = "6.3"
numexpr = "2.8.3"
numpy = "1.22.3"
numpy-base = "1.22.3"
oauth2client = "4.1.3"
openjdk = "11.0.13"
openssl = "1.1.1q"
packaging = "21.3"
pandas = "1.4.3"
paramiko = "2.8.1"
parso = "0.8.3"
patsy = "0.5.2"
pbzip2 = "1.1.13"
pcre = "8.45"
peppy = "0.32.0"
perl = "5.26.2"
pexpect = "4.8.0"
pickleshare = "0.7.5"
pigz = "2.6"
pip = "22.1.2"
plac = "1.3.4"
pluggy = "1.0.0"
ply = "3.11"
prettytable = "3.3.0"
prompt-toolkit = "3.0.20"
protobuf = "3.20.1"
psutil = "5.9.0"
ptyprocess = "0.7.0"
pulp = "2.6.0"
pure_eval = "0.2.2"
py = "1.11.0"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycparser = "2.21"
pygments = "2.11.2"
pynacl = "1.4.0"
pyopenssl = "22.0.0"
pyparsing = "3.0.4"
pyrsistent = "0.18.0"
pysam = "0.18.0"
pysamstats = "1.1.2"
pysftp = "0.2.9"
pysocks = "1.7.1"
pytables = "3.7.0"
pytest = "7.1.2"
python-dateutil = "2.8.2"
python-fastjsonschema = "2.15.1"
python-irodsclient = "1.1.4"
python-isal = "0.11.1"
python_abi = "3.9"
pytz = "2022.1"
pyyaml = "6.0"
ratelimiter = "1.2.0"
readline = "8.1.2"
requests = "2.28.1"
retry = "0.9.2"
rich = "10.16.2"
rsa = "4.7.2"
s3transfer = "0.6.0"
sambamba = "0.8.2"
samclip = "0.4.0"
samtools = "1.15.1"
scipy = "1.8.1"
setuptools = "61.2.0"
six = "1.16.0"
slacker = "0.14.0"
smart_open = "5.2.1"
smmap = "4.0.0"
snakemake = "7.9.0"
snakemake-minimal = "7.9.0"
sqlite = "3.38.5"
stack_data = "0.2.0"
statsmodels = "0.13.2"
stone = "3.2.1"
stopit = "1.1.2"
tabulate = "0.8.10"
tbb = "2020.3"
tk = "8.6.12"
tomli = "2.0.1"
toposort = "1.7"
traitlets = "5.1.1"
typing-extensions = "4.1.1"
typing_extensions = "4.1.1"
tzdata = "2022a"
ubiquerg = "0.6.2"
uritemplate = "4.1.1"
urllib3 = "1.26.9"
veracitools = "0.1.3"
wcwidth = "0.2.5"
wheel = "0.37.1"
wrapt = "1.14.1"
xopen = "1.5.0"
xz = "5.2.5"
yaml = "0.2.5"
yarl = "1.6.3"
yte = "1.5.1"
zipp = "3.8.0"
zlib = "1.2.12"
zstd = "1.5.2"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
